+++ projects/core-ui/src/lib/header/header.component.ts
@@ -7,6 +7,8 @@ import { Component, OnInit, Input } from '@angular/core';
 })
 export class HeaderComponent implements OnInit {
   @Input('ukis-title') title: string = '';
+  @Input('ukis-short-title') shortTitle: string = '';
+  @Input('ukis-version') version: string = '';
   constructor() { }
 
   ngOnInit() {