+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug Fixes
 * **@dlr-eoc/core-ui:**
+  - Fix Styles for `clr-vertical-nav.right` trigger margin.
   - Fix Styles for `clr-vertical-nav.right` icon direction.
 
 # [8.0.1](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v8.0.1) (2022-01-05) (map-ol bug and PR template)